% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVdagger-package.R
\docType{package}
\name{PVdagger-package}
\alias{PVdagger}
\alias{PVdagger-package}
\title{PVdagger: Visualization of Causal Structures in Pharmacovigilance Data Using DAGs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The PVdagger package provides tools for creating and visualizing Directed Acyclic Graphs (DAGs) with various biases and paths. This package is particularly useful for researchers and signal managers in pharmacovigilance who need to illustrate causal relationships and potential biases in their data. The package leverages the DiagrammeR library to generate high-quality DAGs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pvverse.github.io/pv_dagger/}
}

}
\author{
\strong{Maintainer}: Michele Fusaroli \email{michele.fusaroli@who-umc.org}

Authors:
\itemize{
  \item Joseph Mitchell
  \item Annette Rudolph
  \item Elena Rocca
  \item Riccardo Fusaroli
}

}
\keyword{internal}
